+++ src/app/api/boards/route.ts
@@ -1,0 +1,49 @@
+import { NextRequest, NextResponse } from "next/server";
+import { prisma } from "@/lib/prisma";
+
+export async function GET(req: NextRequest) {
+  const boardId = Number(req.nextUrl.searchParams.get("boardId"));
+  if (isNaN(boardId)) {
+    return NextResponse.json(
+      { message: "boardId가 필요합니다" },
+      { status: 400 }
+    );
+  }
+
+  const posts = await prisma.post.findMany({
+    where: {
+      boardId,
+      deletedAt: null,
+    },
+    orderBy: { createdAt: "desc" },
+    include: { profile: true, board: true },
+  });
+
+  return NextResponse.json(posts);
+}
+
+export async function POST(req: NextRequest) {
+  const { title, content, boardId } = await req.json();
+
+  const profile = await prisma.profile.findFirst(); // 테스트용 임시 사용자
+  const board = await prisma.board.findUnique({ where: { boardId } });
+
+  if (!profile || !board) {
+    return NextResponse.json(
+      { message: "profile 또는 board가 존재하지 않습니다." },
+      { status: 400 }
+    );
+  }
+
+  const post = await prisma.post.create({
+    data: {
+      title,
+      content,
+      boardId,
+      profileId: profile.profileId,
+      nickName: profile.nickName,
+    },
+  });
+
+  return NextResponse.json(post, { status: 201 });
+}
